Renamed existing plant images to follow new naming convention
Now that I have some images for the plant maturity, I added the maturity
to the name of the image between the plant name and the health. This
makes more sense because the maturity is more important than the health
of the plant.

I also changed the way plant images are chosen to display the correct
plant image